Refactor SpotsScrollView to be shared between tvOS and iOS
