+++ Sources/iOS/Classes/SpotsScrollView.swift
@@ -3,6 +3,8 @@ import QuartzCore
 
 /// The core foundation scroll view inside of Spots that manages the linear layout of all components.
 open class SpotsScrollView: UIScrollView, UIGestureRecognizerDelegate {
+  var sizeCache = [Int: CGFloat]()
+
   private struct Observer: Equatable {
     let view: UIView
     let keyValueObservation: NSKeyValueObservation
@@ -13,7 +15,7 @@ open class SpotsScrollView: UIScrollView, UIGestureRecognizerDelegate {
   }
 
   /// A collection of UIView's that resemble the order of the views in the scroll view
-  fileprivate var subviewsInLayoutOrder = [UIView]()
+  var subviewsInLayoutOrder = [UIView]()
   private var observers = [Observer]()
 
   /// The distance that the content view is inset from the enclosing scroll view.
@@ -166,74 +168,6 @@ open class SpotsScrollView: UIScrollView, UIGestureRecognizerDelegate {
       })
 
       observers.append(Observer(view: view, keyValueObservation: boundsObserver))
-    }
-  }
-
-  /// Layout views in linear order based of view index in `subviewsInLayoutOrder`
-  func layoutViews() {
-    guard let superview = superview else {
-      return
-    }
-
-    componentsView.frame = bounds
-    componentsView.bounds = CGRect(origin: contentOffset, size: bounds.size)
-
-    var yOffsetOfCurrentSubview: CGFloat = 0.0
-    let lastView = subviewsInLayoutOrder.last
-
-    for subview in subviewsInLayoutOrder {
-      if let scrollView = subview as? UIScrollView {
-        var frame = scrollView.frame
-        var contentOffset = scrollView.contentOffset
-
-        if self.contentOffset.y < yOffsetOfCurrentSubview {
-          contentOffset.y = 0.0
-          frame.origin.y = yOffsetOfCurrentSubview
-        } else {
-          contentOffset.y = self.contentOffset.y - yOffsetOfCurrentSubview
-          frame.origin.y = self.contentOffset.y
-        }
-
-        let remainingBoundsHeight = fmax(bounds.maxY - frame.minY, 0.0)
-        let remainingContentHeight = fmax(scrollView.contentSize.height - contentOffset.y, 0.0)
-
-        if configuration.stretchLastComponent && scrollView.isEqual(lastView) {
-          let newHeight = self.frame.size.height - scrollView.frame.origin.y + self.contentOffset.y
-          frame.size.height = newHeight
-        } else {
-          frame.size.height = ceil(fmin(remainingBoundsHeight, remainingContentHeight))
-        }
-
-        // Using `.integral` can sometimes set the height back to 1.
-        // To avoid this we check if the height is zero before we run `.integral`.
-        // If it was, then we set it to zero again to not have frame heights jump between
-        // one and zero when scrolling. Jump frame heights can cause rendering issues and
-        // make `UICollectionView` not render corretly when you use multiple components.
-        let shouldResetFrameHeightToZero = frame.size.height == 0
-        frame = frame.integral
-        if shouldResetFrameHeightToZero {
-          frame.size.height = 0
-        }
-
-        scrollView.frame = frame
-        scrollView.contentOffset = CGPoint(x: Int(contentOffset.x), y: Int(contentOffset.y))
-
-        yOffsetOfCurrentSubview += scrollView.contentSize.height
-      } else {
-        var frame = subview.frame
-        frame.origin.x = 0
-        frame.origin.y = yOffsetOfCurrentSubview
-        subview.frame = frame
-
-        yOffsetOfCurrentSubview += frame.size.height
-      }
-    }
-
-    let minimumContentHeight = bounds.height - (contentInset.top + contentInset.bottom)
-    contentSize = CGSize(width: bounds.size.width, height: fmax(yOffsetOfCurrentSubview, minimumContentHeight))
-
-    if self.frame.size.height != superview.frame.size.height {
-      self.frame.size.height = superview.frame.size.height
     }
   }
